new-text model updated
