+++ testing/text-testing.txt
@@ -1,1 +1,4 @@
-Hi, I’m [Name], 45 years old. I’ve been working in [profession/industry] for quite a while, and I really enjoy [something you like about your work or life]. Over the years, I’ve learned the value of patience, consistency, and staying curious. Outside of work, I love [hobbies/interests—like traveling, gardening, sports, or spending time with family]. I believe every stage of life brings something new to learn, and I enjoy connecting with people and sharing experiences
+Hi, I’m venki, 45 years old. I’ve been working in cts for quite a while, and I really enjoy 
+driving bike. Over the years, I’ve learnedthe value of patience, consistency, and staying 
+curious. Outside of work, I love spending time with family. I believe every stage of life 
+brings something new to learn, and I enjoy connecting with people and sharing experiences
